Refactor argument parser and config manager
